Added new \Common\Core\Init class to use inheritance for Backend, Frontend and Api Init classes.
